extract input processing out of ray caster
